% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_cells.R
\name{downsample_cells}
\alias{downsample_cells}
\title{Downsample to get fewer cells from the specified cell type}
\usage{
downsample_cells(keptExp, prop_cell = NULL, n_top_genes = NULL)
}
\arguments{
\item{keptExp}{A dgCMatrix object where genes as rows and cells as
columns. Output from scFlowExamples::merge_zeisel_celltypes() function.}

\item{prop_cell}{A vector containing cell proportion to downsample.
Should be of same length as the cell types in the dgCMatrix. Value
ranges between 0-1.}

\item{n_top_genes}{Number of top variable genes to downsample. If NULL
all genes are returned.}
}
\description{
Downsample to get fewer cells from the specified cell type
}
\author{
Nurun Fancy \href{mailto:n.fancy@imperial.ac.uk}{n.fancy@imperial.ac.uk}
}
