% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_cells.R
\name{ds_cell}
\alias{ds_cell}
\title{Downsampling function for a single cell type}
\usage{
ds_cell(exp, prop_cell = prop_cell)
}
\arguments{
\item{exp}{A dgCMatrix object where genes as rows and cells as
columns.}

\item{prop_cell}{Proportion of original cell number to downsample.
Value ranges between 0-1. Providing expected proportion will override
n_cell.}

\item{cell_type}{Cell type name to downsample from.}

\item{n_cell}{Number of cells required after downsampling.}
}
\description{
Downsampling function for a single cell type
}
\author{
Nurun Fancy <n.fancy@imperial.ac.uk>
}
