% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_symbol_to_human_ensembl.R
\name{mouse_symbol_to_human_ensembl}
\alias{mouse_symbol_to_human_ensembl}
\title{Convert gene symbols from mouse to human and then to human ensembl IDs}
\usage{
mouse_symbol_to_human_ensembl(dt)
}
\arguments{
\item{dt}{Object of class dgCMatrix}
}
\description{
Convert gene symbols from mouse to human and then to human ensembl IDs
}
