% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseSymbol_to_humanEnsembl.r
\name{mouseSymbol_to_humanEnsembl}
\alias{mouseSymbol_to_humanEnsembl}
\title{Convert gene symbols from mouse to human and then to human ensembl IDs}
\usage{
mouseSymbol_to_humanEnsembl(x, ...)
}
\arguments{
\item{x}{Object of class dgCMatrix}

\item{...}{additional parameters}
}
\description{
Convert gene symbols from mouse to human and then to human ensembl IDs
}
