% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_zeisel2018.r
\name{prep_zeisel2018}
\alias{prep_zeisel2018}
\title{Prep the zeisel2018 dataset so that it is in sparse matrix format}
\usage{
prep_zeisel2018(path)
}
\arguments{
\item{path}{Path to the loom file}
}
\value{
A list containing $exp with the sparse matrix and $annot with annot data
}
\description{
\code{prep_zeisel2018} Loads the Zeisel2018 dataset in Loom format and converts to sparse matrix
}
\examples{
# create a temporary directory
td = tempdir()
 create the placeholder file
tf = tempfile(tmpdir=td, fileext=".zip")
download into the placeholder file
download.file("https://storage.googleapis.com/linnarsson-lab-loom/l5_all.loom", tf)
zeisel2018 = prep_zeisel2018()
}
