% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_celltypes_byIndv.r
\name{split_celltypes_byIndv}
\alias{split_celltypes_byIndv}
\title{#########################################################################
Split the celltypes by assinging individual ids.}
\usage{
split_celltypes_byIndv(
  keptExp,
  joinCells = c("TEGLU1", "TEGLU2"),
  nCases = 3,
  jointName = "Pyramidal"
)
}
\arguments{
\item{keptExp}{A list containing all the experiment samples}

\item{joinCells}{The celltypes to join as a single type of cells which are
then split into two groups of samples. i.e. cases and controls}

\item{nCases}{Number of individual of cases to split}

\item{jointName}{New name of the joined celltype}
}
\value{
Returns an `outputList` which is what format and information?
}
\description{
#########################################################################
Split the celltypes by assinging individual ids.
}
