% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scflowexample.R
\name{write_scflow_manifest}
\alias{write_scflow_manifest}
\title{Write Manifest file to be run by scFlow and nf-core-scflow}
\usage{
write_scflow_manifest(dataset = "indvExp", output_dir = getwd())
}
\arguments{
\item{dataset}{the dataset to be used to write to the location from scFlowExamples. Default is indvExp, other choices indvExp_ds and indvExp_ds_4K}

\item{output_dir}{Path for the output dir. Default is current directory.}
}
\description{
Write Manifest file to be run by scFlow and nf-core-scflow
}
\author{
Nurun Fancy \href{mailto:n.fancy@imperial.ac.uk}{n.fancy@imperial.ac.uk}
}
