% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scflowexample.R
\name{write_scflow_samplesheet}
\alias{write_scflow_samplesheet}
\title{Write SampleSheet file to be run by scFlow and nf-core-scflow}
\usage{
write_scflow_samplesheet(indvExp = indvExp, output_dir = ".")
}
\arguments{
\item{indvExp}{indvExp loaded by the first function write_data.}

\item{output_dir}{Path for the output dir. Default is current directory.}
}
\description{
The scFlowExample dataset, Manifest.txt and SampleSheet.tsv files can
be prepared using three simple functions.
}
\author{
Nurun Fancy <n.fancy@imperial.ac.uk>
}
