% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scflowexample.R
\name{write_data}
\alias{write_data}
\title{Generate scFlowExample dataset for scFlow and nf-core-scflow}
\usage{
write_data(output_dir = ".")
}
\arguments{
\item{output_dir}{Path for the output dir. Default is current directory}
}
\description{
Generate scFlowExample dataset for scFlow and nf-core-scflow
}
\author{
Nurun Fancy <n.fancy@imperial.ac.uk>
}
