% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scflowexample.R
\name{write_scflow_manifest}
\alias{write_scflow_manifest}
\title{Write Manifest file to be run by scFlow and nf-core-scflow}
\usage{
write_scflow_manifest(indvExp = indvExp, output_dir = getwd())
}
\arguments{
\item{indvExp}{indvExp loaded by the first function write_data.}

\item{output_dir}{Path for the output dir. Default is current directory.}
}
\description{
Write Manifest file to be run by scFlow and nf-core-scflow
}
\author{
Nurun Fancy <n.fancy@imperial.ac.uk>
}
