% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scflowexample.R
\name{write_scflow_samplesheet}
\alias{write_scflow_samplesheet}
\title{Write SampleSheet file to be run by scFlow and nf-core-scflow}
\usage{
write_scflow_samplesheet(dataset = "indvExp", output_dir = getwd())
}
\arguments{
\item{dataset}{the dataset to be used to write to the location from scFlowExamples. Default is indvExp, other choices indvExp_ds and indvExp_ds_4K}

\item{output_dir}{Path for the output dir. Default is current directory.}
}
\description{
The scFlowExample dataset, Manifest.txt and SampleSheet.tsv files can
be prepared using three simple functions.
}
\author{
Nurun Fancy \href{mailto:n.fancy@imperial.ac.uk}{n.fancy@imperial.ac.uk}
}
